Rename `avoidCapture` to `getAvailableVariableName`
