+++ rules/prefer-array-find.js
@@ -9,7 +9,7 @@ import {
 	isLeftHandSide,
 	singular,
 	getScopes,
-	avoidCapture,
+	getAvailableVariableName,
 	getVariableIdentifiers,
 } from './utils/index.js';
 import {isMethodCall} from './ast/index.js';
@@ -316,7 +316,7 @@ const create = context => {
 				const singularName = singular(node.id.name);
 				if (singularName) {
 					// Rename variable to be singularized now that it refers to a single item in the array instead of the entire array.
-					const singularizedName = avoidCapture(singularName, getScopes(scope));
+					const singularizedName = getAvailableVariableName(singularName, getScopes(scope));
 					yield * renameVariable(variable, singularizedName, fixer);
 
 					// Prevent possible variable conflicts
