+++ rules/utils/index.js
@@ -21,10 +21,12 @@ export {
 	getBooleanAncestor,
 } from './boolean.js';
 
-export {default as avoidCapture} from './avoid-capture.js';
+export {default as cartesianProductSamples} from './cartesian-product-samples.js';
 export {default as escapeString} from './escape-string.js';
+export {default as getAvailableVariableName} from './get-available-variable-name.js';
 export {default as getCallExpressionArgumentsText} from './get-call-expression-arguments-text.js';
 export {default as getCallExpressionTokens} from './get-call-expression-tokens.js';
+export {default as getIndentString} from './get-indent-string.js';
 export {default as getReferences} from './get-references.js';
 export {default as getScopes} from './get-scopes.js';
 export {default as getVariableIdentifiers} from './get-variable-identifiers.js';
@@ -39,11 +41,14 @@ export {default as isOnSameLine} from './is-on-same-line.js';
 export {default as isSameIdentifier} from './is-same-identifier.js';
 export {default as isSameReference} from './is-same-reference.js';
 export {default as isShadowed} from './is-shadowed.js';
+export {default as isShorthandImportLocal} from './is-shorthand-import-local.js';
+export {default as isShorthandPropertyValue} from './is-shorthand-property-value.js';
 export {default as isValueNotUsable} from './is-value-not-usable.js';
 export {default as needsSemicolon} from './needs-semicolon.js';
-export {default as shouldAddParenthesesToMemberExpressionObject} from './should-add-parentheses-to-member-expression-object.js';
-export {default as shouldAddParenthesesToCallExpressionCallee} from './should-add-parentheses-to-call-expression-callee.js';
 export {default as shouldAddParenthesesToAwaitExpressionArgument} from './should-add-parentheses-to-await-expression-argument.js';
+export {default as shouldAddParenthesesToCallExpressionCallee} from './should-add-parentheses-to-call-expression-callee.js';
+export {default as shouldAddParenthesesToConditionalExpressionChild} from './should-add-parentheses-to-conditional-expression-child.js';
+export {default as shouldAddParenthesesToMemberExpressionObject} from './should-add-parentheses-to-member-expression-object.js';
 export {default as singular} from './singular.js';
 export {default as toLocation} from './to-location.js';
 export {default as getAncestor} from './get-ancestor.js';
